Corrected bugs when cleaning systems
