+++ src/map/include/systems/updatesystem.h
@@ -6,7 +6,7 @@ namespace Systems {
 
 class UpdateSystem : public System {
     public:
-        UpdateSystem(SystemManager&) = default;
+        UpdateSystem(SystemManager&) {}
         virtual ~UpdateSystem() = default;
 
         virtual void update(EntityManager &es, double);
